+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change log
 
-## future release
+## 2.7.0
 
 * Change Home Assistant auto-discovery log level to `debug` (https://github.com/Jalle19/eda-modbus-bridge/issues/93)
 * Add settings switches for cooling/heating allowed (https://github.com/Jalle19/eda-modbus-bridge/issues/98)